requirements.txt
# These are the requirements needed for python

mysql-connector-python
flask